upload Genpos and PEP Rel 2.3.1.012 builds
provide new builds of Rel 2.3.1, Rel 2.3.1.012, which contains a fix for a defect in the PLU Maintenance dialog's Control Settings for Remote Printers #5 through #8 along with biometrics library removed.

The Control Settings dialog of the PLU Maintenance dialog had a defect in which the check boxes for Remote Printers #5 through #8 always indicated the printers were not 
being used even when the bits indicating those printers were in use were set. So the Remote Printer settings for Remote Printers #5 through #8 did not accurately reflect
the actual settings for those devices.