Rel 2.3.1.13 with modifications for SQLite ODBC
These are the build products from a new branch of Rel 2.3.1 with the following changes:
 - connection string now uses SQLite3 ODBC driver for PLU totals database
 - minor changes to PEP to remove pack to one byte compiler option for VS 2019

The last build of Rel 2.3.1 was with Visual Studio 2005. In order to provide a build for testing a connection string using SQLite3 ODBC driver for the PLU totals database rather than the SQL Server, we had to port the source to Visual Studio 2019. Several of the projects were using a compiler option to pack structs to a one byte boundary. This caused a compilation error in a Windows include which detected the option as being set and triggered a compilation error.